+++ test/web/core.js
@@ -201,6 +201,27 @@ success
 }
 */
 
+// allcaps methods
+
+done = false;
+startTime = Date.now();
+var res = local.dispatch({
+  POST: 'httpl://test.com/foo', Accept: 'text/plain',
+  body: 'echo this, please', Content_Type: 'text/plain'
+});
+res.then(printSuccess, printError).always(finishTest);
+wait(function () { return done; });
+
+/* =>
+success
+{
+  body: "echo this, please",
+  headers: {"content-type": "text/plain"},
+  reason: "ok",
+  status: 200
+}
+*/
+
 // uppercase headers in responses
 
 done = false;